refactot: rearrange folders to pages & components
