+++ src/components/Tabs/ArrowDownBtn.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../../utils";
+import { cn } from "../../utils";
 interface ArrowDownBtnProps {
   onMoveDown: (tabId: number, direction: "down") => void;
   tabId: number;
