+++ src/pages/Home/CopyToClipboard.tsx
