+++ src/pages/index.tsx
@@ -1,0 +1,2 @@
+export { default as Home } from "./Home";
+export { default as TrackTime } from "./TrackTime";
