+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type { Space, SpaceDoc } from "./space";
+export type { Tab } from "./tab";
+export type { UrlDuration } from "./webtime";
