+++ src/types/webtime.ts
@@ -1,0 +1,9 @@
+import { FieldValue } from "firebase/firestore";
+export interface UrlDuration {
+  id: string;
+  url: string;
+  faviconUrl: string;
+  durationBySecond: number;
+  lastVisitTime: FieldValue;
+  visitCounts: number;
+}
